requirements.txt
certifi==2021.5.30
influxdb-client==1.18.0
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
Rx==3.2.0
six==1.16.0
urllib3==1.26.5
